+++ src/widgets/qownnotesmarkdowntextedit.h
@@ -31,7 +31,8 @@ class QOwnNotesMarkdownTextEdit : public QMarkdownTextEdit {
     void updateSettings();
     QMargins viewportMargins();
     void setText(const QString &text);
-    static void setSpellcheckingEnabled(bool enabled);
+    static void setSpellCheckingEnabled(bool enabled);
+    bool isSpellCheckingEnabled();
 
    protected:
     // we must not override _highlighter or Windows will create a